Modernize the web parts of the project.
We now use simple-routes, allowing for much easier maintenance of the
project down the road.

In passing, optimize the generation of the RSS bits by writing down a
stream piecemeal rather than bubbling the whole thing into memory first.
On my local machine, not doing that meant visiting the LDB...