Import a freshly compiled website.
The compilation process doesn't take care of static resources as of yet,
so I did a manual sync of the images/ and static/ subdirs. Let's publish
and see what's missing now --- as we use "/static/styles.css"
references, it's hard to guess from the local files version (to browse
it, I typically run a local Common Lisp webserver that does dynamic
rendering of the website for me, so from the master's branch).

This gh-pages workflow will need some taking care of.