Try to adapt to gh-pages setup.
The static content is not to be found at the / location but at the
/tapoueh.org/ location, apparently. That's because I chose to make
tapoueh.org its own project's gh-pages for publishing rather than use
the per-user setup of gh-pages. But I don't want to revisit that choice,
I want the blog to fully work at https://dimitri.github.io/tapoueh.org/.

Current guess is that more work is going to be needed, but we're back to
commit, push and see workflow here it seems: is there a way to run a
test instance of gh-pages setup on my laptop?