Switch (back?) to relative URLs.
Lots of efforts went into the sotfware to have “beautiful” URLs used
everywhere, assuming publication on a top-level directory of a domain, such
as at http://tapoueh.org.

To be able to easily switch to any other hosting solution we need to have
the site fully workable on-disk as pure static content, that is a set of
relocatable files using only (self-contained) relative URLs.

This patch is very much WIP towards that goal. See #1.