Some more fixes...
The idea is still to be able to publish again the content, under any hosting
facility, which means compiling a fully relocatable website. It used not to
be the case.

Let's see how much github pages like this use of symlinks, the previous try
apparently had some leftovers that might have been hard to swallow...

See #1.