Inverse symlinks and storage.
Let's see if GH pages prefers it this way, now we have all the static
content and images stored into the /docs part of the reposiroty and symlinks
are used from the root of it so that our files continue to make sense both
at the source level (muse) and the built level (html).

Also symlinks are used when browsing the local dynamic website.

See #1 once again.