New version of tapoueh.org.
This version uses https://gohugo.io/ static website generator, because in
2017 at least we have public maintained tools to publish a modern good
looking entirely static website.

This is the first commit of the migration to Hugo, with the generated files
in the docs/ directory as before, so that we can publish directly on github.

The migration script is available in content/post/migration.lisp.

The articles are now written in Markdown.