Review migration script: <center> and skipped content.
In the Muse based publication system that we used to use, the first image of
the article, located after the first paragraph, would be used as a cover
image. Our Muse parser allows us to easily turn that image into a proper
cover image in hugo terms.

The problem is that if we skip this image from the contents of the article
then we might display an orphan comment on the image. Add the smarts to the
migration script for it to know about the situation and then skip the manual
centered legend that goes with the skipped image that is now a cover.

Some other desultory changes.