Review missing content, avoid 404.
We need some more aliases before to consider going live with the new version
of the site, so here we add them. Not all of them, but almost. Note that the
missing content was found with:

  diff -qr tapoueh.org.muse/docs/ tapoueh.org/docs/ \
  | grep "Only in tapoueh.org.muse"                 \
  | egrep -v '/tags:|/images:|/confs:|index.html|archives.html|cloud.json|static|thumbnails'

After all when switching from a static website only publishing solution to
another, all it takes is comparing the distribution .html files on disk in
order to find potential 404s...