Review H1/H2 situation, one H1 per page is all we need.
In passing, also edit the look of the website to some extend, because we had
to edit the templates anyway. The idea is still to optimize SEO, in
particular of the last year's articles.

In this patch we also display only 5 articles per listing page: the
summaries have gotten big enough to warrant shortening the pages from 10
article summaries down to only 5.